+++ resources/views/partials/dividers/archive/curve.blade.php
@@ -5,6 +5,6 @@
 -->
 
 <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-viewBox="0 0 1920.19 95.98" preserveAspectRatio="none" class="divider <?php echo $classes; ?>">
-	 <path d="M1920.1.5c-337.23,66.13-664.31,93.25-960,95C664.41,93.75,337.33,66.63.1.5Z"/>
+	 viewBox="0 0 1200 120" preserveAspectRatio="none" class="divider <?php echo $classes; ?>">
+	 <path d="M1200,0v64.9c-133.4,28.5-352.5,47.1-600,47.1S133.4,93.4,0,64.9V0H1200z"/>
 </svg>