+++ resources/views/partials/dividers/curve_invert.blade.php
@@ -6,5 +6,6 @@
 
 <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
 viewBox="0 0 1920.19 95.98" preserveAspectRatio="none" class="divider <?php echo $classes; ?>">
-	 <path d="M1920.1.5c-337.23,66.13-664.31,93.25-960,95C664.41,93.75,337.33,66.63.1.5Z"/>
+<path d="M1920.5.61h-960c295.69,1.73,622.77,28.85,960,95Z"/>
+<path d="M960.5.61H.5v95C337.73,29.46,664.81,2.34,960.5.61Z"/>
 </svg>